Make estimations its own file
